Move two functions to helpers
