+++ tests/unit/cli/admin/metadata/test_sign.py
@@ -4,13 +4,12 @@
 
 import json
 
-import click
 import pretend
-import pytest
 
 from repository_service_tuf.cli.admin.metadata import sign
-from repository_service_tuf.helpers.api_client import URL, Methods
+from repository_service_tuf.helpers.api_client import URL
 from tests.conftest import _HELPERS, _PAYLOADS, _PEMS, _ROOTS, invoke_command
+
 
 
 class TestSign:
@@ -29,14 +28,10 @@ class TestSign:
         )
 
         with open(f"{_PAYLOADS / 'sign_pending_roles.json'}") as f:
-            fake_response_data = json.load(f)
+            full_data = json.load(f)
 
-        fake_response = pretend.stub(
-            json=pretend.call_recorder(lambda: fake_response_data),
-            status_code=200,
-        )
-        sign.request_server = pretend.call_recorder(
-            lambda *a, **kw: fake_response
+        sign._get_pending_roles = pretend.call_recorder(
+            lambda *a, **kw:  full_data["data"]["metadata"]
         )
         sign.send_payload = pretend.call_recorder(lambda **kw: "fake-taskid")
         sign.task_status = pretend.call_recorder(lambda *a: "OK")
@@ -54,10 +49,8 @@ class TestSign:
             result.data["signature"]["keyid"] == expected["signature"]["keyid"]
         )
         assert "Metadata Signed and sent to the API! 🔑" in result.stdout
-        assert sign.request_server.calls == [
-            pretend.call(
-                api_server, "api/v1/metadata/sign/", Methods.GET, headers=None
-            )
+        assert sign._get_pending_roles.calls == [
+            pretend.call(test_context["settings"])
         ]
         assert sign.send_payload.calls == [
             pretend.call(
@@ -103,15 +96,8 @@ class TestSign:
         with open(f"{_ROOTS / 'v1.json'}") as f:
             v1_das_root = f.read()
 
-        fake_response_data = {
-            "data": {"metadata": {"root": json.loads(v1_das_root)}}
-        }
-        fake_response = pretend.stub(
-            json=pretend.call_recorder(lambda: fake_response_data),
-            status_code=200,
-        )
-        sign.request_server = pretend.call_recorder(
-            lambda *a, **kw: fake_response
+        sign._get_pending_roles = pretend.call_recorder(
+            lambda *a, **kw: {"root": json.loads(v1_das_root)}
         )
         sign.send_payload = pretend.call_recorder(lambda **kw: "fake-taskid")
         sign.task_status = pretend.call_recorder(lambda *a: "OK")
@@ -128,10 +114,8 @@ class TestSign:
         assert result.data["role"] == "root"
         assert result.data["signature"]["keyid"] == expected["keyid"]
         assert "Metadata Signed and sent to the API! 🔑" in result.stdout
-        assert sign.request_server.calls == [
-            pretend.call(
-                api_server, "api/v1/metadata/sign/", Methods.GET, headers=None
-            )
+        assert sign._get_pending_roles.calls == [
+            pretend.call(test_context["settings"])
         ]
         assert sign.send_payload.calls == [
             pretend.call(
@@ -327,19 +311,8 @@ class TestSignError:
         with open(f"{_ROOTS / 'v2.json'}") as f:
             v2_das_root = f.read()
 
-        fake_response_data = {
-            "data": {
-                "metadata": {
-                    "root": json.loads(v2_das_root),
-                }
-            }
-        }
-        fake_response = pretend.stub(
-            json=pretend.call_recorder(lambda: fake_response_data),
-            status_code=200,
-        )
-        sign.request_server = pretend.call_recorder(
-            lambda *a, **kw: fake_response
+        sign._get_pending_roles = pretend.call_recorder(
+            lambda *a, **kw: {"root": json.loads(v2_das_root)}
         )
         api_server = "http://127.0.0.1"
         test_context["settings"].SERVER = api_server
@@ -355,11 +328,9 @@ class TestSignError:
         )
 
         assert test_result.exit_code == 1, test_result.stdout
-        assert "Previous root v1 needed to sign root v2" in test_result.output
-        assert sign.request_server.calls == [
-            pretend.call(
-                api_server, "api/v1/metadata/sign/", Methods.GET, headers=None
-            )
+        assert "Previous root v1 needed to sign root v2" in test_result.stdout
+        assert sign._get_pending_roles.calls == [
+            pretend.call(test_context["settings"])
         ]
 
     def test_sign_fully_signed_metadata(
@@ -371,19 +342,8 @@ class TestSignError:
         with open("tests/files/payload/ceremony.json", "r") as f:
             ceremony_payload = json.loads(f.read())
 
-        fake_response_data = {
-            "data": {
-                "metadata": {
-                    "root": ceremony_payload["metadata"]["root"],
-                }
-            }
-        }
-        fake_response = pretend.stub(
-            json=pretend.call_recorder(lambda: fake_response_data),
-            status_code=200,
-        )
-        sign.request_server = pretend.call_recorder(
-            lambda *a, **kw: fake_response
+        sign._get_pending_roles = pretend.call_recorder(
+            lambda *a, **kw: {"root": ceremony_payload["metadata"]["root"]}
         )
         api_server = "http://127.0.0.1"
         test_context["settings"].SERVER = api_server
@@ -399,70 +359,7 @@ class TestSignError:
         )
 
         assert test_result.exit_code == 1, test_result.stdout
-        assert "Metadata already fully signed." in test_result.output
-        assert sign.request_server.calls == [
-            pretend.call(
-                api_server, "api/v1/metadata/sign/", Methods.GET, headers=None
-            )
-        ]
-
-
-class TestHelpers:
-    def test__parse_pending_data(self):
-        fake_md = ["md1", "md2"]
-        result = sign._parse_pending_data({"data": {"metadata": fake_md}})
-
-        assert result == fake_md
-
-    def test__parse_pending_data_missing_metadata(self):
-        with pytest.raises(click.ClickException) as e:
-            sign._parse_pending_data({"data": {}})
-
-        assert "No metadata available for signing" in str(e)
-
-    def test__get_pending_roles_request(self, monkeypatch):
-        fake_settings = pretend.stub(SERVER=None, HEADERS=None)
-        fake_json = pretend.stub()
-        response = pretend.stub(
-            status_code=200, json=pretend.call_recorder(lambda: fake_json)
-        )
-        sign.request_server = pretend.call_recorder(lambda *a, **kw: response)
-
-        parsed_data = pretend.stub()
-        fake__parse_pending_data = pretend.call_recorder(lambda a: parsed_data)
-        monkeypatch.setattr(
-            sign, "_parse_pending_data", fake__parse_pending_data
-        )
-
-        result = sign._get_pending_roles(fake_settings)
-
-        assert result == parsed_data
-        assert sign.request_server.calls == [
-            pretend.call(
-                fake_settings.SERVER,
-                sign.URL.METADATA_SIGN.value,
-                sign.Methods.GET,
-                headers=None,
-            )
-        ]
-        assert response.json.calls == [pretend.call()]
-        assert fake__parse_pending_data.calls == [pretend.call(fake_json)]
-
-    def test__get_pending_roles_request_bad_status_code(self):
-        fake_settings = pretend.stub(
-            SERVER="http://localhost:80", HEADERS=None
-        )
-        response = pretend.stub(status_code=400, text="")
-        sign.request_server = pretend.call_recorder(lambda *a, **kw: response)
-        with pytest.raises(click.ClickException) as e:
-            sign._get_pending_roles(fake_settings)
-
-        assert "Failed to fetch metadata for signing" in str(e)
-        assert sign.request_server.calls == [
-            pretend.call(
-                fake_settings.SERVER,
-                sign.URL.METADATA_SIGN.value,
-                sign.Methods.GET,
-                headers=None,
-            )
+        assert "Metadata already fully signed." in test_result.stdout
+        assert sign._get_pending_roles.calls == [
+            pretend.call(test_context["settings"])
         ]
