+++ tests/unit/cli/admin/metadata/test_update.py
